implement hal::PwmPin
